execution cache test cleanup
